fix: Allow bal-addresses to ingest the new balancer-deployments directory structure.
This approach takes a minimal change approach, whereby in the resulting addresbook the v2/v3 directory name is ignored/not present. It assumes that all v3 deployments will start with v3- and we can filter on this.  This approach invovles no braking changes to anything that currently runs, but also provides the least structure/information about what belongs to what.